data_ingestion/requirements.txt
requests==2.31.0
six
pyspark==3.3.1
kafka-python
sentence-transformers
langdetect
torch
transformers
tqdm
mlflow
apscheduler